Revert "Fix debug text on some emulators"
This reverts commit 063ed6b54316e294cbcfcee916b72106e0953193.

+++ tools/sotn-debugmode/debugmode.h
@@ -1,6 +1,6 @@
 #include <game.h>
 
 void InitFont();
-void SetFontCoord(s16 x, s16 y);
+void SetFontCoord(int x, int y);
 void BeginFont();
 void EndFont();
